Move LocalFileWatcher to linux directory, as it is specific for Linux
